pyproject.toml
[project]
name = "golf-reservation-chatbot"
version = "0.1.0"
description = "A golf course reservation chatbot powered by MCP and OpenAI"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0,<0.116.0",
    "uvicorn[standard]>=0.23.0",
    "openai>=1.30.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
mcp = [
    "mcp[cli]>=1.1.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
